Fix intensity numerical issue, fix tests

+++ test_sais.jl
@@ -3,4 +3,4 @@ include("toy_normal.jl")
 include("simple_mixture.jl")
 
 backend = CUDABackend()
-ais(SimpleMixture(backend); elt_type = Float64, backend)
+ais(SimpleMixture(backend); elt_type = Float32, backend)
